feat: реализация Basic Auth, проверки прав доступа

+++ src/main/java/ru/skypro/homework/service/AdService.java
@@ -2,6 +2,7 @@ package ru.skypro.homework.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.skypro.homework.dto.ad.AdDto;
@@ -11,6 +12,7 @@ import ru.skypro.homework.dto.ad.ExtendedAd;
 import ru.skypro.homework.mapper.AdMapper;
 import ru.skypro.homework.model.Ad;
 import ru.skypro.homework.repository.AdRepository;
+import ru.skypro.homework.util.SecurityUtils;
 
 import java.util.List;
 
@@ -20,50 +22,49 @@ import java.util.List;
 public class AdService {
 
     private final AdRepository adRepository;
-
     private final AdMapper adMapper;
-
+    private final UserService userService;
 
     public AdsDto getListAd() {
         List<Ad> adList = adRepository.findAll();
-        List<AdDto> adDto = adList.stream()
-                .map(adMapper::toDto)
-                .toList();
+        List<AdDto> adDto = adList.stream().map(adMapper::toDto).toList();
         return new AdsDto(adDto.size(), adDto);
     }
 
     @Transactional
-    public AdDto createAd(CreateOrUpdateAd dto) { //TODO проверка создателя и добавление его в объявление
+    public AdDto createAd(CreateOrUpdateAd dto, UserDetails userDetails) {
 
         Ad ad = adMapper.toEntity(dto);
+        ad.setAuthor(userService.checkUser(userDetails.getUsername()));
         adRepository.save(ad);
 
         return adMapper.toDto(ad);
     }
 
     public ExtendedAd getAdBuId(Integer id) {
-
-        return adMapper.toExtendedDto(adRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Неправильный идентификатор комментария")));
+        return adMapper.toExtendedDto(adRepository.findById(id).
+                orElseThrow(() -> new EntityNotFoundException("Неправильный идентификатор комментария")));
     }
 
     @Transactional
-    public void deleteAd(Integer id) {
+    public void deleteAd(Integer id, UserDetails userDetails) {
         Ad ad = checkAd(id);
+        SecurityUtils.checkModifyPermission(ad.getAuthor(), userDetails);
         adRepository.delete(ad);
     }
 
     @Transactional
-    public AdDto updateAd(Integer id, CreateOrUpdateAd dto) {
+    public AdDto updateAd(Integer id, CreateOrUpdateAd dto, UserDetails userDetails) {
         Ad ad = checkAd(id);
+        SecurityUtils.checkModifyPermission(ad.getAuthor(), userDetails);
         adMapper.updateAd(dto, ad);
         return adMapper.toDto(ad);
+
     }
 
-    public AdsDto getListAdUserAuth() {  //TODO доделать когда будет готова авторизация
-        List<Ad> adList = adRepository.findAll();
-        List<AdDto> adDto = adList.stream()
-                .map(adMapper::toDto)
-                .toList();
+    public AdsDto getListAdUserAuth(UserDetails userDetails) {
+        List<Ad> adList = adRepository.findAllByAuthorEmail(userDetails.getUsername());
+        List<AdDto> adDto = adList.stream().map(adMapper::toDto).toList();
         return new AdsDto(adDto.size(), adDto);
     }
 
@@ -71,4 +72,5 @@ public class AdService {
     private Ad checkAd(Integer id) {
         return adRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Неправильный идентификатор комментария"));
     }
+
 }
