+++ src/main/java/ru/skypro/homework/service/CommentService.java
@@ -2,6 +2,7 @@ package ru.skypro.homework.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.skypro.homework.dto.comment.CommentDto;
@@ -11,6 +12,7 @@ import ru.skypro.homework.mapper.CommentMapper;
 import ru.skypro.homework.model.Comment;
 import ru.skypro.homework.repository.AdRepository;
 import ru.skypro.homework.repository.CommentRepository;
+import ru.skypro.homework.util.SecurityUtils;
 
 import java.util.List;
 
@@ -23,11 +25,11 @@ public class CommentService {
     private final CommentMapper commentMapper;
     private final AdRepository adRepository;
 
+    private final UserService userService;
+
     public CommentsDto getListComments(Integer idAd) {
         checkAdBuId(idAd);
-
         List<Comment> commentList = commentRepository.findAllByAd_Id(idAd);
-
         List<CommentDto> commentsDto = commentList.stream()
                 .map(commentMapper::toDto)
                 .toList();
@@ -36,31 +38,37 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentDto createComment(Integer idAd, CreateOrUpdateComment dto) { //TODO пользователя берем из авторизации
+    public CommentDto createComment(Integer idAd, CreateOrUpdateComment dto, UserDetails userDetails) {
         checkAdBuId(idAd);
 
         Comment comment = commentMapper.toEntity(dto);
         comment.setId(idAd);
+        comment.setAuthor(userService.checkUser(userDetails.getUsername()));
         commentRepository.save(comment);
         return commentMapper.toDto(comment);
     }
 
     @Transactional
-    public CommentDto updateComment(Integer idAd, Integer commentId, CreateOrUpdateComment dto) {
-        commentRepository.findById(commentId).orElseThrow(()-> new EntityNotFoundException("Неправильный идентификатор комментария"));
+    public CommentDto updateComment(Integer idAd, Integer commentId, CreateOrUpdateComment dto, UserDetails userDetails) {
+        commentRepository.findById(commentId).
+                orElseThrow(() -> new EntityNotFoundException("Неправильный идентификатор комментария"));
         checkAdBuId(idAd);
+
         Comment comment = commentMapper.toEntity(dto);
+        SecurityUtils.checkModifyPermission(comment.getAuthor(), userDetails);
         comment.setText(dto.text());
         commentRepository.save(comment);
         return commentMapper.toDto(comment);
     }
 
     @Transactional
-    public void deleteComment (Integer idAd, Integer idComment) {
-        Comment comment = commentRepository.findById(idComment).orElseThrow(()-> new EntityNotFoundException("Неправильный идентификатор комментария"));
-        if (comment.getAd().equals(idAd)){
+    public void deleteComment(Integer idAd, Integer idComment, UserDetails userDetails) {
+        Comment comment = commentRepository.findById(idComment).
+                orElseThrow(() -> new EntityNotFoundException("Неправильный идентификатор комментария"));
+        if (comment.getAd().equals(idAd)) {
             throw new EntityNotFoundException("Неправильный идентификатор объявления");
         }
+        SecurityUtils.checkModifyPermission(comment.getAuthor(), userDetails);
         commentRepository.delete(comment);
     }
